+++ src/server.rs
@@ -1,0 +1,5 @@
+use crate::command::Command;
+
+pub trait Server<'a> {
+    fn navigate(&self, command: Command) -> Result<(), Box<dyn std::error::Error>>;
+}
